skills and goals moved into /assets/datas
