+++ src/App.js
@@ -5,8 +5,8 @@ import Header from "./components/Header";
 import Section from "./common/Section";
 import Gallery from "./features/repos/Gallery";
 import Footer from "./components/Footer";
-import skills from "./skills";
-import goals from "./goals";
+import skills from "./assets/datas/skills";
+import goals from "./assets/datas/goals";
 import { lightMode, darkMode } from './theme';
 import { useDispatch, useSelector } from "react-redux";
 import { fetchThemeState, selectTheme } from "./features/themes/themeSlice";
