+++ src/assets/datas/skills.js
